+++ .next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"khEathkoXprds1kv0npp3yQW8UJsmxs/mMG22Hz2LGE=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"hHAoGNHutJE2CJZlkEJFHROeF02BtLhjPGo16bXk9EI=\"\n}"