+++ .next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "3b031f0b4f96a26a"; }
+/******/ 		__webpack_require__.h = function() { return "4b7a5e67775a5bd3"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */